remove ui helper
