+++ src/main.js
@@ -5,8 +5,6 @@ import store from "./store";
 
 import './styles/main.css'
 
-console.log('hi')
-
 Vue.config.productionTip = false;
 
 new Vue({
